+++ bsp_tool/branches/arkane/dark_messiah_singleplayer.py
@@ -14,7 +14,7 @@ BSP_VERSION = (20, 4)  # int.from_bytes(struct.pack("2H", 20, 4), "little")
 
 GAME_PATHS = ["Dark Messiah of Might and Magic Single Player"]
 
-GAME_VERSIONS = {GAME: BSP_VERSION for GAME in GAME_PATHS}
+GAME_VERSIONS = {GAME_PATH: BSP_VERSION for GAME_PATH in GAME_PATHS}
 
 
 LUMP = orange_box.LUMP