+++ bsp_tool/branches/infinity_ward/call_of_duty1.py
@@ -14,7 +14,7 @@ BSP_VERSION = 59
 
 GAME_PATHS = ["Call of Duty", "Call of Duty: United Offensive"]
 
-GAME_VERSIONS = {GAME: BSP_VERSION for GAME in GAME_PATHS}
+GAME_VERSIONS = {GAME_PATH: BSP_VERSION for GAME_PATH in GAME_PATHS}
 
 
 class LUMP(enum.Enum):