+++ bsp_tool/branches/infinity_ward/call_of_duty4.py
@@ -13,7 +13,7 @@ BSP_VERSION = 22
 
 GAME_PATHS = ["Call of Duty 4: Modern Warfare"]
 
-GAME_VERSIONS = {GAME: BSP_VERSION for GAME in GAME_PATHS}
+GAME_VERSIONS = {GAME_PATH: BSP_VERSION for GAME_PATH in GAME_PATHS}
 
 
 # NOTE: lumps are given ids and headers reference these ids in order