+++ bsp_tool/branches/valve/goldsrc.py
@@ -24,7 +24,7 @@ GAME_PATHS = [*[f"Half-Life/{mod}" for mod in [
                             "valve"]],   # Half-Life
               "Halfquake Trilogy", "Sven Co-op"]
 
-GAME_VERSIONS = {GAME: BSP_VERSION for GAME in GAME_PATHS}
+GAME_VERSIONS = {GAME_PATH: BSP_VERSION for GAME_PATH in GAME_PATHS}
 
 
 class LUMP(enum.Enum):