+++ bsp_tool/branches/valve/sdk_2013.py
@@ -2,6 +2,9 @@
 import enum
 import struct
 
+# from . import alien_swarm
+# from . import left4dead
+# from . import left4dead2
 from . import orange_box
 from . import source
 
@@ -10,13 +13,85 @@ FILE_MAGIC = b"VBSP"
 
 BSP_VERSION = 21
 
-GAME_PATHS = ["Blade Symphony", "Counter-Strike: Global Offensive", "Portal 2",
-              "Source Filmmaker"]
+GAME_PATHS = ["Blade Symphony/berimbau",
+              "Counter-Strike Global Offensive/csgo",
+              "infra/infra",
+              "Portal 2/portal2",
+              "Source Filmmaker/game/tf"]
+# also sourcemods / mapbase
 
-GAME_VERSIONS = {GAME: BSP_VERSION for GAME in GAME_PATHS}
+GAME_VERSIONS = {"infra/infra": 22,
+                 **{GAME_PATH: BSP_VERSION for GAME_PATH in GAME_PATHS if GAME_PATH != "infra/infra"}}
 
 
-LUMP = orange_box.LUMP
+# Counter-Strike Global Offensive/bin/bsppack.dll
+class LUMP(enum.Enum):
+    ENTITIES = 0
+    PLANES = 1
+    TEXTURE_DATA = 2
+    VERTICES = 3
+    VISIBILITY = 4
+    NODES = 5
+    TEXTURE_INFO = 6
+    FACES = 7
+    LIGHTING = 8
+    OCCLUSION = 9
+    LEAVES = 10
+    FACE_IDS = 11
+    EDGES = 12
+    SURFEDGES = 13
+    MODELS = 14
+    WORLD_LIGHTS = 15
+    LEAF_FACES = 16
+    LEAF_BRUSHES = 17
+    BRUSHES = 18
+    BRUSH_SIDES = 19
+    AREAS = 20
+    AREA_PORTALS = 21
+    FACE_BRUSHES = 22  # infra
+    FACE_BRUSH_LIST = 23  # infra
+    UNUSED_24 = 24
+    UNUSED_25 = 25
+    DISPLACEMENT_INFO = 26
+    ORIGINAL_FACES = 27
+    PHYSICS_DISPLACEMENT = 28
+    PHYSICS_COLLIDE = 29
+    VERTEX_NORMALS = 30
+    VERTEX_NORMAL_INDICES = 31
+    DISPLACEMENT_LIGHTMAP_ALPHAS = 32  # deprecated / X360 ?
+    DISPLACEMENT_VERTICES = 33
+    DISPLACEMENT_LIGHTMAP_SAMPLE_POSITIONS = 34
+    GAME_LUMP = 35
+    LEAF_WATER_DATA = 36
+    PRIMITIVES = 37
+    PRIMITIVE_VERTICES = 38  # deprecated / X360 ?
+    PRIMITIVE_INDICES = 39
+    PAKFILE = 40
+    CLIP_PORTAL_VERTICES = 41
+    CUBEMAPS = 42
+    TEXTURE_DATA_STRING_DATA = 43
+    TEXTURE_DATA_STRING_TABLE = 44
+    OVERLAYS = 45
+    LEAF_MIN_DIST_TO_WATER = 46
+    FACE_MACRO_TEXTURE_INFO = 47
+    DISPLACEMENT_TRIS = 48
+    PROP_BLOB = 49  # left4dead
+    WATER_OVERLAYS = 50  # deprecated / X360 ?
+    LEAF_AMBIENT_INDEX_HDR = 51
+    LEAF_AMBIENT_INDEX = 52
+    LIGHTING_HDR = 53
+    WORLD_LIGHTS_HDR = 54
+    LEAF_AMBIENT_LIGHTING_HDR = 55
+    LEAF_AMBIENT_LIGHTING = 56
+    XZIP_PAKFILE = 57  # deprecated / X360 ?
+    FACES_HDR = 58
+    MAP_FLAGS = 59
+    OVERLAY_FADES = 60
+    OVERLAY_SYSTEM_LEVELS = 61  # left4dead
+    PHYSICS_LEVEL = 62  # left4dead2
+    DISPLACEMENT_MULTIBLEND = 63  # alienswarm
+
+# TODO: Known lump changes from Orange Box -> Source SDK 2013:
 
 
 # struct SourceBspHeader { char file_magic[4]; int version; SourceLumpHeader headers[64]; int revision; };