+++ tests/test_d3d_bsp.py
@@ -9,19 +9,20 @@ from bsp_tool.branches.infinity_ward import call_of_duty4
 
 d3dbsps = []
 # TODO: add more Call of Duty 4 dirs from maplist.installed_games & make it optional
-for map_dir in ["./tests/maps/Call of Duty 4", "./tests/maps/Call of Duty 4/mp"]:
-    for map_name in fnmatch.filter(os.listdir(map_dir), ".d3dbsp"):
+for map_dir in ["tests/maps/Call of Duty 4", "tests/maps/Call of Duty 4/mp"]:
+    map_dir = os.path.join(os.getcwd(), map_dir)
+    for map_name in fnmatch.filter(os.listdir(map_dir), "*[Bb][Ss][Pp]"):
         d3dbsps.append(D3DBsp(call_of_duty4, os.path.join(map_dir, map_name)))
 
 
 @pytest.mark.parametrize("d3dbsp", d3dbsps)
 def test_no_errors(d3dbsp):
-    assert len(d3dbsp.loading_errors) == 0
+    assert len(d3dbsp.loading_errors) == 0, d3dbsp.filename
 
 
 @pytest.mark.parametrize("d3dbsp", d3dbsps)
 def test_entites_loaded(d3dbsp):
-    assert d3dbsp.ENTITIES[0]["classname"] == "worldspawn"
+    assert d3dbsp.ENTITIES[0]["classname"] == "worldspawn", d3dbsp.filename
 
 
 # TODO: implement .save_as method and test that uneditted saves match EXACTLY