+++ tests/test_idtech_bsp.py
@@ -8,9 +8,9 @@ from bsp_tool import IdTechBsp
 from bsp_tool.branches.id_software import quake3
 
 bsps = []
-map_dir = "./tests/maps/Quake 3 Arena"
+map_dir = os.path.join(os.getcwd(), "tests/maps/Quake 3 Arena")
 # TODO: add more Quake 3 Arena dirs from maplist.installed_games & make it optional
-for map_name in fnmatch.filter(os.listdir(map_dir), ".d3dbsp"):
+for map_name in fnmatch.filter(os.listdir(map_dir), "*[Bb][Ss][Pp]"):
     bsps.append(IdTechBsp(quake3, os.path.join(map_dir, map_name)))
 
 
@@ -21,7 +21,7 @@ def test_no_errors(bsp: IdTechBsp):
 
 @pytest.mark.parametrize("bsp", bsps)
 def test_entities_loaded(bsp: IdTechBsp):
-    assert bsp.ENTITIES[0]["classname"] == "worldspawn"
+    assert bsp.ENTITIES[0]["classname"] == "worldspawn", bsp.filename
 
 
 # NOTE: WIP
@@ -29,7 +29,7 @@ def test_entities_loaded(bsp: IdTechBsp):
 def test_face_struct(bsp: IdTechBsp):  # the most complex MappedArray
     # TODO: add more asserts, be thorough
     header = bsp.headers["FACES"]
-    assert header.length % struct.calcsize(quake3.Face._format) == 0
+    assert header.length % struct.calcsize(quake3.Face._format) == 0, bsp.filename
     with open(bsp.file.name, "rb") as file:
         file.seek(header.offset)
         raw_faces = file.read(header.length)