+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.4.0 (???)
+## v0.4.0 (??? 2022)
 
 ## New
  * Added support for Ritual Entertainment's Ubertools (Quake III Engine Branch)
@@ -18,17 +18,25 @@
  * `load_bsp` now only accepts a `branch_script` as it's optional argument
 
 ### Newly Supported
+ * Infinity Ward Engine
+  - IW 2.0
+  - IW 3.0
+ * Ion Storm IdTech
+  - Daikatana
+ * Raven IdTech
  * Source Engine
    - Tactical Intervention
    - Vampire the Masquerade: Bloodlines
  * Ubertools
 
 ### Updated Support
- * Quake Engine
-   - Hexen 2
  * Id Tech 3
    - Quake III Arena
    - Raven Software Titles
+ * Infinity Ward Engine
+   - Call of Duty
+ * Quake Engine
+   - Hexen 2
  * Source Engine
  * Titanfall Engine